KAFKA-5017: Record batch first offset remains accurate after compaction

+++ clients/src/main/java/org/apache/kafka/common/record/DefaultRecord.java
@@ -103,7 +103,7 @@ public class DefaultRecord implements Record {
     }
 
     @Override
-    public long sequence() {
+    public int sequence() {
         return sequence;
     }
